--- src/main/kotlin/com/zup/pagamento/model/FormaPagamentoResponse.kt
@@ -1,7 +1,0 @@
-package com.zup.pagamento.model
-
-interface FormaPagamentoResponse {
-    val id: Long?
-
-    val descricao: String?
-}
