+++ src/main/kotlin/com/zup/pagamento/service/RegraFraude.kt
@@ -3,6 +3,6 @@ package com.zup.pagamento.service
 import com.zup.pagamento.dto.UsuarioDto
 
 interface RegraFraude {
-    fun avaliarFormaPagamentoDisponivelParaUsuarios(): List<UsuarioDto>
-    fun retornarUsuariosFraudadores(): List<String>
+    fun avaliarFormaPagamentoDisponivelParaUsuario(idRestaurante: Long, email: String): List<UsuarioDto>
+    fun retornarUsuariosFraudadores(): Set<String>
 }
