+++ src/main/kotlin/com/zup/pagamento/service/RegraFraudeImp.kt
@@ -1,6 +1,7 @@
 package com.zup.pagamento.service
 
 import com.zup.pagamento.dto.UsuarioDto
+import com.zup.pagamento.repository.FormaPagamentoRepository
 import com.zup.pagamento.repository.FraudadorRepository
 import com.zup.pagamento.repository.UsuarioRepository
 import org.springframework.beans.factory.annotation.Autowired
@@ -15,20 +16,21 @@ class RegraFraudeImp : RegraFraude {
     @Autowired
     lateinit var usuarioRepository: UsuarioRepository
 
-    override fun avaliarFormaPagamentoDisponivelParaUsuarios(): List<UsuarioDto> {
-        var fraudadores = retornarUsuariosFraudadores()
+    override fun avaliarFormaPagamentoDisponivelParaUsuario(idRestaurante: Long, email: String): List<UsuarioDto> {
+        val fraudadores = retornarUsuariosFraudadores()
         var us2: List<UsuarioDto> = listOf()
 
-        usuarioRepository.findAll().forEach {
-            if (fraudadores.contains(it.email)) {
-                it.formasPagamento = it.formasPagamento.filter { fm -> !fm.tipo.pagamentoOnline }.toSet()
-            }
-            us2 += listOf<UsuarioDto>(it.converterResponse())
-        }
+        val formasPagamentoUsuario = usuarioRepository.formasPagamentoPorUsuarioERestauranteId(idRestaurante, email)
 
+        for (usuario in formasPagamentoUsuario) {
+            if (fraudadores.contains(usuario.email)) {
+                usuario.formasPagamento = usuario.formasPagamento.filter { fm -> !fm.tipo.pagamentoOnline }.toSet()
+                us2 += listOf<UsuarioDto>(usuario.converterResponse())
+            } else us2 = listOf<UsuarioDto>(usuario.converterResponse())
+        }
         return us2
     }
 
-    override fun retornarUsuariosFraudadores(): List<String> = fraudadorRepository.findAll().map { it.email }
+    override fun retornarUsuariosFraudadores(): Set<String> = fraudadorRepository.findAll().map { it.email }.toSet()
 
 }
